Add GitHub configuration
This patch adds two GitHub Actions—one for building the package with a
compiler matrix and one for building the sdist and ensuring it
works—as well as configuration for GitHub’s Dependabot and GitHub’s
sponsorship program.  This configuration is based on the configuration
in the `github:pniedzielski/rdf4h-vocab-activitystreams` repository.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>
License: BSD-3-Clause